requirements.txt
google-generativeai>=0.8.5
anthropic>=0.67.0
click>=8.2.1
openai>=1.107.1
PyYAML>=6.0.2
tzlocal>=5.3.1
post_slug>=1.0.1
requests>=2.32.5
filetype>=1.2.0
tqdm>=4.67.1
colorama>=0.4.6

# Test dependencies
pytest>=8.4.2
pytest-cov>=7.0.0
